CoreformIGA

CoreformIGA.BasisBernstein.B

CoreformIGA.BasisBernstein.dBdxi

CoreformIGA.BasisBernstein.plot!

CoreformIGA.BasisBernstein.B — Method

B(p, xi)

Return the value of the Bernstein basis of degree $p$ evaluated at $\xi \in [0,1]$.

CoreformIGA.BasisBernstein.dBdxi — Method

dBdxi(p, xi)

Return the first derivative of the Bernstein basis of degree $p$ evaluated at $\xi \in [0,1]$.

See also: B(p, xi)

CoreformIGA.BasisBernstein.plot! — Method

plot!(plt, p, f; subd=10)

Plot the values or first derivatives (as indicated by the passed in function B(p, xi) or dBdxi(p, xi)) of the Bernstein basis of degree $p$ over $[0,1]$.

See also: B(p, xi), dBdxi(p, xi)