CoreformIGA

CoreformIGA.ContinuousComponent

CoreformIGA.ContinuousComponent.FunctionCollection

CoreformIGA.BasisBernstein.B

CoreformIGA.BasisBernstein.B1d

CoreformIGA.BasisBernstein.dBdxi

CoreformIGA.BasisBernstein.dBdxi1d

CoreformIGA.BasisBernstein.dnBdxin

CoreformIGA.BasisBernstein.dnBdxin1d

CoreformIGA.BasisBernstein.B — Method

B(p, xi)

Return the value of the tensor product Bernstein basis of degree $p$ evaluated at $\xi$. This can be multidimensional and so both $p$ and $n$ are arrays to specify the basis order in each direction and the number of derivatives in each direction, respectively.

CoreformIGA.BasisBernstein.B1d — Method

B1d(p, xi)

Return the value of the 1d Bernstein basis of degree $p$ evaluated at $\xi \in [0,1]$.

CoreformIGA.BasisBernstein.dBdxi — Method

dBdxi(p,xi)

Return the gradient of the tensor product Bernstein basis of degree $p$ evaluated at $\xi$. This can be multidimensional and so both $p$ and $n$ are arrays to specify the basis order in each direction and the number of derivatives in each direction, respectively.

CoreformIGA.BasisBernstein.dBdxi1d — Method

dBdxi1d(p, xi)

Return the first derivative of the 1d Bernstein basis of degree $p$ evaluated at $\xi \in [0,1]$.

See also: B1d(p, xi)

CoreformIGA.BasisBernstein.dnBdxin — Method

dnBdxin(p, n, xi)

Return the $nth$ derivative of the tensor product Bernstein basis of degree $p$ evaluated at $\xi$. This can be multidimensional and so both $p$ and $n$ are arrays to specify the basis order in each direction and the number of derivatives in each direction, respectively.

CoreformIGA.BasisBernstein.dnBdxin1d — Method

dnBdxin1d(p, n, xi)

Return the nth derivative of the 1d Bernstein basis of degree $p$ evaluated at $\xi \in [0,1]$.

See also: dBdxi1d(p, xi)

CoreformIGA.ContinuousComponent — Module

A container to hold a function collection which includes all function spaces and fields used to conduct the integral of boundary conditions or stiffness matrices, etc.

CoreformIGA.ContinuousComponent.FunctionCollection — Type

Members:

integral: specify the integrand and quadrature information;

geom_field_fc: offer geometry info. for the inverse mapping and integral weights, etc., in the assembly;

test_index_fc: store the DOFs indexes and related functionalities for the test function space;

trial_index_fc: store the DOFs indexes and related functionalities for the trial function space;

test_fs_fc: test function space for evaluating spline functions and indexing them;